requirements.txt
streamlit>=1.37.0
langgraph>=0.2.16
pydantic>=2.8.2
openai>=1.46.0
httpx>=0.27.0
genanki>=0.13.1
python-docx>=1.1.2
pymupdf>=1.24.9
rapidfuzz>=3.9.6
numpy>=1.26.4
pandas>=2.2.2
python-dotenv>=1.0.1
